third_party/WebKit/LayoutTests/editing/selection/after-removing-textarea.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  'before<textarea>|</textarea>after',
  selection => {
    selection.document.querySelector('textarea').remove();
  },
  'before|after'),
  'We should have caret after removing TEXTAREA');
</script>
